changed amount datatype to FLOAT

+++ db_init/init.sql
@@ -16,7 +16,7 @@ CREATE TABLE cryptHubSchema.wallet (
   wallet_id SERIAL PRIMARY KEY,
   user_id INT REFERENCES cryptHubSchema.users(id),
   currency VARCHAR(50),
-  amount DECIMAL(10, 2),
+  amount FLOAT,
   CONSTRAINT unique_user_currency UNIQUE (user_id, currency)
 );
